+++ translator/src/main/resources/lib/soda/lib/Comb.scala
@@ -17,7 +17,7 @@ trait CartesianProduct {
         else
           {
             lazy val rev_sequences = sequences.reverse
-            Rec () .foldLeft (rev_sequences.tail, initial_value (rev_sequences.head ), next_value ) }
+            Rec () .fold (rev_sequences.tail, initial_value (rev_sequences.head ), next_value ) }
 
       def initial_value (seq: Seq [T]  ): Seq [Seq [T]] = seq.map (elem => Seq (elem )  )