+++ translator/src/main/resources/lib/soda/lib/Recursive.scala
@@ -9,7 +9,7 @@ package soda.lib
  */
 trait Recursive {
 
-  def foldLeftWhile [A, B, C <: B]  (s: Seq [A], initial_value: C, next_value: (B, A ) => C, condition: (B, A ) => Boolean ): C =
+  def fold [A, B, C <: B]  (sequence: Seq [A], initial_value: C, next_value_function: (B, A ) => C, condition: (B, A ) => Boolean ): C =
     {
       import scala.annotation.tailrec
         @tailrec
@@ -19,22 +19,22 @@ trait Recursive {
         else
           if (! condition (acc, seq.head )
           ) acc
-          else rec (seq.tail, next_value (acc, seq.head )  )
+          else rec (seq.tail, next_value_function (acc, seq.head )  )
 
-      rec (s, initial_value ) }
+      rec (sequence, initial_value ) }
 
-  def foldLeft [A, B, C <: B]  (seq: Seq [A], initial_value: C, next_value: (B, A ) => C ): C =
+  def fold [A, B, C <: B]  (sequence: Seq [A], initial_value: C, next_value_function: (B, A ) => C ): C =
     {
       import scala.annotation.tailrec
         @tailrec
       def rec (seq: Seq [A], acc: C ): C =
         if (seq.isEmpty
         ) acc
-        else rec (seq.tail, next_value (acc, seq.head )  )
+        else rec (seq.tail, next_value_function (acc, seq.head )  )
 
-      rec (seq, initial_value ) }
+      rec (sequence, initial_value ) }
 
-  def range (n: Int ): Seq [Int] =
+  def range (size: Int ): Seq [Int] =
     {
       import scala.annotation.tailrec
         @tailrec
@@ -43,7 +43,7 @@ trait Recursive {
         ) seq
         else rec (n - 1, seq.+: (n - 1 )  )
 
-      rec (n, Seq [Int]  ()  ) }
+      rec (size, Seq [Int]  ()  ) }
 }
 
 case class Rec () extends Recursive