+++ translator/src/main/scala/soda/translator/io/SimpleIO.scala
@@ -38,7 +38,7 @@ trait SimpleIO {
       .map (x => writer.flush () )
       .map (x => writer.close () )
       .map (x => true )
-      .get
+      .value
 
   def create_file (parent_directory: String, file_name: String ): File =
     new File (parent_directory, file_name )