+++ translator/src/main/scala/soda/translator/language/LineJoiner.scala
@@ -66,7 +66,7 @@ trait Joiner {
     ) lines_to_join
     else
       {
-        lazy val tuple = Rec () .foldLeft (lines_to_join.tail, _initial_value (lines_to_join.head ), _next_value )
+        lazy val tuple = Rec () .fold (lines_to_join.tail, _initial_value (lines_to_join.head ), _next_value )
         lazy val result =
           if (tuple.in_process_rev.isEmpty
           ) tuple.processed_rev.+: (tuple.previous_line )